tests passing with split pipelines
