+++ .env.example
@@ -5,7 +5,7 @@ S3_ADDRESS=minio
 S3_PORT=9000
 S3_SECRET_KEY=minioadmin
 S3_ACCESS_KEY=minioadmin
-S3_DEFAULT_BUCKET=gleanerbucket
+S3_DEFAULT_BUCKET=iow
 S3_USE_SSL=false
 S3_REGION=us
 TRIPLESTORE_URL=http://graphdb:7200
